+++ tests/run-server.sh
@@ -2,16 +2,21 @@
 
 set -e
 
-# Add /.local to path
-export PATH=$PATH:/.local/bin
-
 echo "-- HOME is $HOME"
 
-echo "Installing python package, please wait...."
+VENV_PATH=/.local/venv
 
-pip3 install -U -q --user setuptools
-pip3 install --no-warn-script-location -q --user --prefer-binary -r requirements.txt 
-pip3 install --user -e ./ 
+PIP="$VENV_PATH/bin/pip"
+PIP_INSTALL="$VENV_PATH/bin/pip install -U"
+
+echo "-- Creating virtualenv"
+python3 -m venv --system-site-packages $VENV_PATH
+
+echo "-- Installing required packages..."
+$PIP_INSTALL -q pip setuptools wheel
+$PIP_INSTALL -q --prefer-binary -r requirements.txt
+
+$PIP install -e .
 
 export QGIS_DISABLE_MESSAGE_HOOKS=1
 export QGIS_NO_OVERRIDE_IMPORT=1
@@ -19,5 +24,5 @@ export QGIS_NO_OVERRIDE_IMPORT=1
 export QGSWPS_SERVER_PARALLELPROCESSES=$WORKERS
 
 cd tests/unittests
-exec wpsserver -p 8080
+exec $VENV_PATH/bin/wpsserver -p 8080